feat(transactionDetail) add: transaction detail converter and fix some code

+++ internal/model/transactionDetail.go
@@ -4,7 +4,7 @@ type CreateTransactionDetailRequest struct {
 	TransactionID int64 `json:"transaction_id" validate:"required"`
 	VoucherID     int64 `json:"voucher_id" validate:"required"`
 	CostInPoint   int64 `json:"cost_in_point" validate:"required"`
-	Quantity      int64 `json:"quantity" validate:"required"`
+	Quantity      int   `json:"quantity" validate:"required"`
 	SubTotalCost  int64 `json:"sub_total_cost" validate:"required"`
 }
 
@@ -12,7 +12,7 @@ type CreateTransactionDetailResponse struct {
 	ID            int64  `json:"id"`
 	TransactionID int64  `json:"transaction_id"`
 	VoucherID     int64  `json:"voucher_id"`
-	Quantity      int64  `json:"quantity"`
+	Quantity      int    `json:"quantity"`
 	SubTotalCost  int64  `json:"sub_total_cost"`
 	CreatedAt     string `json:"created_at"`
 }
